+++ examples/create-sst-rds/packages/graphql/genql/runtime/createClient.ts
@@ -1,3 +1,4 @@
+// @ts-nocheck
 
 import  { type BatchOptions, createFetcher } from './fetcher'
 import type { ExecutionResult, LinkedType } from './types'