+++ packages/create-sst/bin/presets/base/javascript/preset.mjs
@@ -3,7 +3,7 @@ import { extract, patch, install } from "create-sst";
 export default [
   extract(),
   install({
-    packages: ["sst", "aws-cdk-lib@2.62.2", "constructs@10.1.156"],
+    packages: ["sst", "aws-cdk-lib@2.72.1", "constructs@10.1.156"],
     dev: true,
   }),
 ];