+++ packages/create-sst/bin/presets/standard/nextjs/preset.mjs
@@ -2,17 +2,23 @@ import { patch, extend, extract, cmd, str_replace } from "create-sst";
 export default [
   extend("presets/base/monorepo"),
   cmd({
-    cmd: "npx create-next-app@latest next --ts --no-eslint --no-src-dir --experimental-app --import-alias '@/*' --use-pnpm",
+    cmd: "npx create-next-app@latest web --ts --no-eslint --no-src-dir --no-experimental-app --import-alias '@/*' --use-pnpm --tailwind",
     cwd: "packages",
   }),
   patch({
-    file: "packages/next/tsconfig.json",
+    file: "packages/web/tsconfig.json",
     operations: [
       {
         op: "add",
         path: "/compilerOptions/paths/@@@app~1core~1*",
         value: ["../core/src/*"],
       },
+    ],
+  }),
+  patch({
+    file: "packages/web/package.json",
+    operations: [
+      { op: "add", path: "/scripts/dev", value: "sst bind next dev" },
     ],
   }),
   extract(),