+++ packages/sst/src/node/future/auth/index.ts
@@ -9,8 +9,11 @@ export * from "./adapter/oidc.js";
 export * from "./adapter/google.js";
 export * from "./adapter/link.js";
 export * from "./adapter/github.js";
+export * from "./adapter/facebook.js";
 export * from "./adapter/oauth.js";
 export type { Adapter } from "./adapter/adapter.js";
 
 export * from "./session.js";
 export * from "./handler.js";
+
+export { Issuer } from "openid-client";