+++ packages/sst/src/util/error.ts
@@ -1,0 +1,6 @@
+export abstract class SSTError extends Error {
+  constructor(message?: string) {
+    super(message);
+    this.name = this.constructor.name;
+  }
+}